+++ src/Application/Operations/Bulk/BulkUpdateOperation.php
@@ -7,14 +7,21 @@ namespace JeroenG\Explorer\Application\Operations\Bulk;
 use JeroenG\Explorer\Application\BePrepared;
 use JeroenG\Explorer\Application\Explored;
 
-class BulkUpdateOperation implements BulkOperationInterface
+final class BulkUpdateOperation implements BulkOperationInterface
 {
     /** @var Explored[] */
     private array $models = [];
 
-    public static function from(iterable $iterable): self
+    private static string $indexName;
+
+    public function __construct(string $indexName)
     {
-        $operation = new self();
+        self::$indexName = $indexName;
+    }
+
+    public static function from(iterable $iterable, string $indexName): self
+    {
+        $operation = new self($indexName);
 
         if ($iterable instanceof \Traversable) {
             $operation->models = iterator_to_array($iterable);
@@ -34,17 +41,17 @@ class BulkUpdateOperation implements BulkOperationInterface
     {
         $payload = [];
         foreach ($this->models as $model) {
-            $payload[] = self::modelToBulkAction($model);
+            $payload[] = self::bulkActionSettings($model);
             $payload[] = self::modelToData($model);
         }
         return $payload;
     }
 
-    private static function modelToBulkAction(Explored $model): array
+    private static function bulkActionSettings(Explored $model): array
     {
         return [
             'index' => [
-                '_index' => $model->searchableAs(),
+                '_index' => self::$indexName,
                 '_id' => $model->getScoutKey(),
             ]
         ];