+++ src/ExplorerServiceProvider.php
@@ -9,12 +9,16 @@ use Illuminate\Foundation\Application;
 use Illuminate\Support\ServiceProvider;
 use JeroenG\Explorer\Application\DocumentAdapterInterface;
 use JeroenG\Explorer\Application\IndexAdapterInterface;
+use JeroenG\Explorer\Application\IndexChangedCheckerInterface;
 use JeroenG\Explorer\Domain\Aggregations\AggregationSyntaxInterface;
 use JeroenG\Explorer\Domain\IndexManagement\IndexConfigurationRepositoryInterface;
 use JeroenG\Explorer\Infrastructure\Console\ElasticSearch;
+use JeroenG\Explorer\Infrastructure\Console\ElasticUpdate;
+use JeroenG\Explorer\Infrastructure\Elastic\ElasticAdapter;
 use JeroenG\Explorer\Infrastructure\Elastic\ElasticClientFactory;
 use JeroenG\Explorer\Infrastructure\Elastic\ElasticDocumentAdapter;
 use JeroenG\Explorer\Infrastructure\Elastic\ElasticIndexAdapter;
+use JeroenG\Explorer\Infrastructure\IndexManagement\ElasticIndexChangedChecker;
 use JeroenG\Explorer\Infrastructure\IndexManagement\ElasticIndexConfigurationRepository;
 use JeroenG\Explorer\Infrastructure\Scout\ElasticEngine;
 use Laravel\Scout\Builder;
@@ -29,16 +33,34 @@ class ExplorerServiceProvider extends ServiceProvider
         }
 
         $this->app->bind(ElasticClientFactory::class, function () {
-            $client = ClientBuilder::create()->setHosts([config('explorer.connection')])->build();
-            return new ElasticClientFactory($client);
+            $client = ClientBuilder::create()->setHosts([config('explorer.connection')]);
+
+            if(config()->has('explorer.connection.api')) {
+                $client->setApiKey(
+                    config('explorer.connection.api.id'),
+                    config('explorer.connection.api.key')
+                );
+            }
+
+            return new ElasticClientFactory($client->build());
         });
 
         $this->app->bind(IndexAdapterInterface::class, ElasticIndexAdapter::class);
 
         $this->app->bind(DocumentAdapterInterface::class, ElasticDocumentAdapter::class);
 
+        $this->app->bind(IndexChangedCheckerInterface::class, ElasticIndexChangedChecker::class);
+
+        $this->app->bind(DeprecatedElasticAdapterInterface::class, function () {
+            $client = ClientBuilder::create()->setHosts([config('explorer.connection')])->build();
+            return new ElasticAdapter($client);
+        });
+
         $this->app->bind(IndexConfigurationRepositoryInterface::class, function () {
-            return new ElasticIndexConfigurationRepository(config('explorer.indexes') ?? []);
+            return new ElasticIndexConfigurationRepository(
+                config('explorer.indexes') ?? [],
+                config('explorer.prune_old_aliases')
+            );
         });
 
         resolve(EngineManager::class)->extend('elastic', function (Application $app) {
@@ -98,6 +120,7 @@ class ExplorerServiceProvider extends ServiceProvider
 
         $this->commands([
              ElasticSearch::class,
+             ElasticUpdate::class,
          ]);
     }
 }