+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - The new match_phrase query (#73)
 - Ability to optionally set more parameters for a few queries (#68)
+- Index aliases, useful for zero downtime indexing
+
+### Changed
+- Connection configuration now can include API ID and key (#74)
 
 ### Deprecated
 - The `elastic:create` and `elastic:delete` are deprecated in favour of `scout:index` and `scout:delete-index`